Cleanup: consolidate the two alloc_apqns_from_mkvp routines
